requirements.txt
pandas
qrcode
pyyaml